Pull up method setTicker(...)
